requirements.txt
Django==6
python-decouple==3.8
psycopg2-binary==2.9.10
Pillow==11.0.0
djangorestframework==3.15.2
django-cors-headers==4.3.1
